Add PolicyServiceFactory class
