+++ components/org.wso2.carbon.identity.api.server.policy/org.wso2.carbon.identity.api.server.policy.v1/src/main/java/org/wso2/carbon/identity/api/server/policy/v1/core/PolicyService.java
@@ -21,7 +21,6 @@ package org.wso2.carbon.identity.api.server.policy.v1.core;
 import org.wso2.carbon.identity.api.server.common.ContextLoader;
 import org.wso2.carbon.identity.api.server.common.Util;
 import org.wso2.carbon.identity.api.server.policy.common.Constants;
-import org.wso2.carbon.identity.api.server.policy.common.PolicyServiceHolder;
 import org.wso2.carbon.identity.api.server.policy.v1.function.PolicyBuilder;
 import org.wso2.carbon.identity.api.server.policy.v1.function.PolicyResponseBuilder;
 import org.wso2.carbon.identity.api.server.policy.v1.model.PolicyListItem;
@@ -47,24 +46,6 @@ public class PolicyService {
 
     private static final int DEFAULT_LIMIT = 30;
     private static final int DEFAULT_OFFSET = 0;
-
-    private static final PolicyService SERVICE;
-
-    static {
-        PolicyManagementService policyManagementService =
-                PolicyServiceHolder.getPolicyManagementService();
-
-        if (policyManagementService == null) {
-            throw new IllegalStateException("PolicyManagementService is not available from OSGi context.");
-        }
-
-        SERVICE = new PolicyService(policyManagementService);
-    }
-
-    public static PolicyService getPolicyService() {
-
-        return SERVICE;
-    }
 
     private final PolicyManagementService policyManagementService;
