+++ components/org.wso2.carbon.identity.api.server.policy/org.wso2.carbon.identity.api.server.policy.v1/src/main/java/org/wso2/carbon/identity/api/server/policy/v1/impl/PoliciesApiServiceImpl.java
@@ -22,6 +22,7 @@ import org.wso2.carbon.identity.api.server.common.ContextLoader;
 import org.wso2.carbon.identity.api.server.policy.common.Constants;
 import org.wso2.carbon.identity.api.server.policy.v1.PoliciesApiService;
 import org.wso2.carbon.identity.api.server.policy.v1.core.PolicyService;
+import org.wso2.carbon.identity.api.server.policy.v1.factories.PolicyServiceFactory;
 import org.wso2.carbon.identity.api.server.policy.v1.model.PolicyRequest;
 import org.wso2.carbon.identity.api.server.policy.v1.model.PolicyResponse;
 import org.wso2.carbon.identity.api.server.policy.v1.model.PolicyUpdateRequest;
@@ -39,7 +40,7 @@ public class PoliciesApiServiceImpl implements PoliciesApiService {
     public PoliciesApiServiceImpl() {
 
         try {
-            this.policyService = PolicyService.getPolicyService();
+            this.policyService = PolicyServiceFactory.getPolicyService();
         } catch (IllegalStateException e) {
             throw new RuntimeException("Error occurred while initiating PolicyService.", e);
         }
